+++ spec/rack/component_spec.rb
@@ -1,92 +1,156 @@
 require 'spec_helper'
-require 'pry'
-require 'securerandom'
 
 RSpec.describe Rack::Component do
-  Fn = proc do |env, &children|
-    <<~HTML
-      <h1>Hello #{env[:name]}</h1>
-      #{children&.call}
-    HTML
-  end
+  describe 'compared to a lambda' do
+    let(:fn) {
+      lambda do |env, &child|
+        <<~HTML
+          <h1>Hello #{CGI.escapeHTML env[:name]}</h1>
+          #{child&.call}
+        HTML
+      end
+    }
+    let(:comp) {
+      Class.new(Rack::Component) do
+        render do |env, &child|
+          <<~HTML
+            <h1>Hello #{h env[:name]}</h1>
+            #{child&.call}
+          HTML
+        end
+      end
+    }
 
-  Comp = Class.new(Rack::Component) do
-    render do |env, &children|
-      <<~HTML
-        <h1>Hello #{env[:name]}</h1>
-        #{children&.call}
-      HTML
-    end
-  end
-
-  describe 'compared to a Proc' do
     it 'behaves identically without children' do
       props = { name: 'Chris' }
-      expect(Comp.call(props)).to eq(Fn.call(props))
+      expect(comp.call(props)).to eq(fn.call(props))
     end
 
     it 'behaves identically with children' do
       props = { name: 'Chris' }
-      expect(Comp.call(props) { 'child' }).to eq(Fn.call(props) { 'child' })
+      actual = comp.call(props) { 'child' }
+      expected = fn.call(props) { 'child' }
+
+      expect(actual).to eq(expected)
+    end
+
+    it 'escapes HTML via a helper method' do
+      result = comp.call(name: '<h1>hi</h1>')
+      expect(result).to include('&lt;h1&gt;hi&lt;/h1&gt;')
     end
   end
 
-  it 'can escape HTML via #h' do
-    comp = Class.new(Rack::Component) { render { |env| h env[:name] } }
-    result = comp.call(name: '<h1>hi</h1>')
-    expect(result).to eq('&lt;h1&gt;hi&lt;/h1&gt;')
-  end
+  describe 'with a block in the +render+ macro' do
+    it 'supports required keywords' do
+      @comp = Class.new(Rack::Component) do
+        render { |name:| name }
+      end
 
-  describe 'handling keyword arguments' do
-    it 'supports required args' do
-      comp = Class.new(Rack::Component) { render { |name:| name } }
-      expect(comp.call(name: 'Jean Luc')).to eq('Jean Luc')
-      expect { comp.call }.to raise_error(ArgumentError)
+      expect(@comp.call(name: 'Jean Luc')).to eq('Jean Luc')
+      expect { @comp.call }.to raise_error(ArgumentError)
     end
 
-    it 'supports optional keyword args' do
-      comp = Class.new(Rack::Component) { render { |name: 'Jean Luc'| name } }
-      expect(comp.call).to eq('Jean Luc')
-      expect(comp.call(name: 'Riker')).to eq('Riker')
+    it 'supports optional keywords' do
+      @comp = Class.new(Rack::Component) do
+        render { |name: 'Jean Luc'| name }
+      end
+
+      expect(@comp.call).to eq('Jean Luc')
+      expect(@comp.call(name: 'Riker')).to eq('Riker')
     end
 
     it 'can set optional keywords with instance methods' do
-      comp = Class.new(Rack::Component) do
+      @comp = Class.new(Rack::Component) do
         render { |name:, dept: department| "#{name} - #{dept}" }
         def department
           'Staff'
         end
       end
-      expect(comp.call(name: 'La Forge')).to eq('La Forge - Staff')
-    end
-  end
-
-  describe 'without a render block' do
-    it 'returns its env by default' do
-      expect(Rack::Component.call).to eq({})
+      expect(@comp.call(name: 'La Forge')).to eq('La Forge - Staff')
     end
 
-    it 'yields self when called with a block' do
-      @comp = Class.new(Rack::Component)
+    it 'can ommit args altogether' do
+      @comp = Class.new(Rack::Component) { render { 'Enterprise' } }
+      expect(@comp.call).to eq('Enterprise')
+    end
 
-      @comp.call(1) { |env| "Hi from comp #{env}" }.tap do |res|
-        expect(res).to eq('Hi from comp 1')
+    it 'allows &block as the only arg' do
+      @comp = Class.new(Rack::Component) do
+        render { |&child| child.call }
       end
+      expect(@comp.call { 'Hi' }).to eq('Hi')
+    end
+
+    it 'can mix keyword args with a block' do
+      @comp = Class.new(Rack::Component) do
+        render do |name:, dept: 'Engineering', **, &child|
+          "#{name} - #{dept} - #{child&.call}"
+        end
+      end
+
+      result = @comp.call(name: 'Barclay') { 'etc' }
+      expect(result).to eq('Barclay - Engineering - etc')
     end
   end
 
-  it 'still works after overriding initialize' do
-    comp = Class.new(Rack::Component) do
+
+  describe 'with a template in the +render+ macro' do
+    let(:expected) { "<h1>Hi, Spock&lt;&gt;</h1>\n" }
+
+    it 'renders ERB' do
+      @comp = Class.new(Rack::Component) do
+        render erb: <<~HTML
+          <h1>Hi, <%= env[:name] %></h1>
+        HTML
+      end
+      expect(@comp.call(name: 'Spock<>')).to eq(expected)
+    end
+
+    it 'renders haml' do
+      @comp = Class.new(Rack::Component) do
+        render haml: <<~'HAML'
+          %h1 Hi, #{env[:name]}
+        HAML
+      end
+      expect(@comp.call(name: 'Spock<>')).to eq(expected)
+    end
+
+     it 'renders the engine you specify' do
+      @comp = Class.new(Rack::Component) do
+        render engine: 'erb', template: <<~ERB
+          <h1>Hi, <%= env[:name] %></h1>
+        ERB
+      end
+      expect(@comp.call(name: 'Spock<>')).to eq(expected)
+     end
+
+     it 'passes template options along via an +opts+ key' do
+       @comp = Class.new(Rack::Component) do
+         render opts: { escape_html: false }, erb: "<%= env[:key] %>"
+       end
+
+       result = @comp.call(key: '<h1>Hi</h1>')
+       expect(result).to eq('<h1>Hi</h1>')
+     end
+
+     it 'supports the +yield+ keyword' do
+       @layout = Class.new(Rack::Component) do
+         render erb: "<%== yield %>"
+       end
+
+       expect(@layout.call { 'Hi' }).to eq('Hi')
+     end
+  end
+
+  it 'fearlessly overrides initialize' do
+    @comp = Class.new(Rack::Component) do
+      render { |_| @id }
+
       def initialize(id)
         @id = id
       end
-
-      render { |env| env }
     end
-    instance = comp.new(1)
-    expect(instance.instance_variable_get(:@env)).to be(nil)
-    expect(instance.instance_variable_get(:@id)).to be(1)
-    expect(instance.call(2)).to eq(2)
+    expect(@comp.call(1)).to eq(1)
   end
 
   it 'renders json swimmingly' do
@@ -97,72 +161,6 @@ RSpec.describe Rack::Component do
 
     @comp.call(captain: 'kirk').tap do |res|
       expect(JSON.parse(res).fetch('captain')).to eq('kirk')
-    end
-  end
-
-  describe 'memoized' do
-    before do
-      @rando = Class.new(Rack::Component) do
-        render { |_| SecureRandom.uuid }
-      end
-    end
-
-    it 'caches identical calls' do
-      @rando.memoized.tap do |uuid|
-        expect(@rando.memoized).to eq(uuid)
-      end
-    end
-
-    it 'works with components that have overriden initialize' do
-      comp = Class.new(Rack::Component) do
-        def initialize(id:, name:)
-          @id = id
-          @name = name
-        end
-
-        render { |_| SecureRandom.uuid }
-      end
-
-      comp.memoized(id: 1, name: "chris").tap do |output|
-        expect(comp.memoized(id: 1, name: "chris")).to eq(output)
-        expect(comp.memoized(id: 2, name: "chris")).not_to eq(output)
-      end
-    end
-
-    it 'limits the cache size to 100 keys by default' do
-      (0..200).map { |key| @rando.memoized(key) }
-      @rando.send(:cache).store.tap do |store|
-        expect(store.length).to eq(100)
-      end
-    end
-
-    it 'overrides the cache size by calling class#cache with a block' do
-      class Tiny < Rack::Component
-        cache { MemoryCache.new(length: 50) }
-        render { |_| "meh" }
-      end
-      (0..200).map { |key| Tiny.memoized(key) }
-      Tiny.send(:cache).store.tap do |store|
-        expect(store.length).to eq(50)
-      end
-    end
-
-    it 'busts cache based on props' do
-      @rando.memoized.tap do |uuid|
-        expect(@rando.memoized(1)).not_to eq(uuid)
-      end
-    end
-
-    it 'does not bust cache based on block' do
-      @rando.memoized.tap do |uuid|
-        expect(@rando.memoized { 'children' }).to eq(uuid)
-      end
-    end
-
-    it 'flushes itself on #flush' do
-      @rando.memoized
-      @rando.flush
-      expect(@rando.send(:cache).store.empty?).to eq(true)
     end
   end
 end