+++ spec/spec_helper.rb
@@ -1,8 +1,8 @@
 require 'bundler/setup'
 require 'rspec'
-require 'rack/component'
-require 'rack/test'
 require 'pry'
+require 'rack/test'
+require 'rack/component'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure